requirements.txt
python-telegram-bot==20.7
qrcode[pil]==7.4.2
python-barcode==0.15.1
Pillow==10.2.0
